Implement player losing one start wanted level when cyberpsycho combat starts.
If players notice cyberpsychos before cyberpsycho combat starts, they
may get a one star wanted level and enter combat with police, which may be an
annoyance to players. This commit makes players lose a one star wanted
level when cyberpsycho combat starts, giving the impression that police
initially attack the player in confusion, but realize the player is
responding to a cyberpsycho attack, or losing interest in the player due
to the larger threat, if the one star wanted level was gained earlier.